+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 #### Fixed
 
-- [#1231](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/1231) Enable identity insert on view's base table
+- [#1232](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/1232) Enable identity insert on view's base table
 
 ## v7.1.7